requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
WTForms==3.1.1
python-dotenv==1.0.0
Werkzeug==3.0.1
bleach==6.1.0
gunicorn==21.2.0
Pillow==10.4.0
